Remove outdated Foundation Safe validation step
This removes a step which validated the nonce change in the Foundation Safe which previously acted as the ProxyAdmin owner. It should have been removed when the runbook was updated for the nested 2 of 2 safe. This new Safe uses a different foundation Safe, so the `0x9ba...` address should not appear at all in the state diff.