Add nonce setting in SIMULATE_WITHOUT_LEDGER branch of PSP ceremony
Fixes an issue where the nonce was not properly set in the tenderly
simulation, when using SIMULATE_WITHOUT_LEDGER.

This occurred because there were two SimulationStateOverrides being set
for the same contract. One in PresignPauseFromJson and the other in
MultisigBuilder._simulateForSigner()
(https://github.com/base-org/contracts/blob/main/script/universal/MultisigBuilder.sol#L134).

Tenderly will ignore the second override, and so the nonce was not being
overriden.
This was probably missed in recent ceremonies because the playbook
developer was also as signer and so did not need to set a TEST_SENDER.