Resign PSPs for DeputyGuardianModule (#220)
* basic setup

* Allow overriding with a task specific PSP solidity script

* Add 008 PSP job to CI config

* Cleanup overrides

* Add draft json files

* Add overrides using changes in base-contracts

* Fix CI

* Document overrides

* Add draft files

* temp changes to base-contracts allowing multiple contract overrides

* Fix test sender addresses to use a mainnet foundation owner

* Fix workdir

* normalize paths

* Move overrides to script in root

* Fix rpc_url

* Add not about deleting wrapper script

* fix ci job name

* Guardian Safe is 1 of 1

Co-authored-by: Matt Solomon <matt@mattsolomon.dev>

* whitespace matters!

Co-authored-by: Matt Solomon <matt@mattsolomon.dev>

* Fix link to paused slot

* Add Safe override documentation

* Add nonce setting in SIMULATE_WITHOUT_LEDGER branch of PSP ceremony

Fixes an issue where the nonce was not properly set in the tenderly
simulation, when using SIMULATE_WITHOUT_LEDGER.

This occurred because there were two SimulationStateOverrides being set
for the same contract. One in PresignPauseFromJson and the other in
MultisigBuilder._simulateForSigner()
(https://github.com/base-org/contracts/blob/main/script/universal/MultisigBuilder.sol#L134).

Tenderly will ignore the second override, and so the nonce was not being
overriden.
This was probably missed in recent ceremonies because the playbook
developer was also as signer and so did not need to set a TEST_SENDER.

* Prevent adding multiple state overrides on the owner safe

* Use _setOverrides in base-contracts

* Remove outdated comment in base-contracts

---------

Co-authored-by: Matt Solomon <matt@mattsolomon.dev>