fp: Dispute Game Blacklist task (#183)
* fp: Dispute Game Blacklist task

Adds a task on Sepolia for blacklisting dispute games from the
`Guardian` safe.

* Add post-checks

* fp: Fall back to `PermissionedDisputeGame` task (#184)

* fp: Fall back to `PermissionedDisputeGame` task

Adds a task for the `Guardian` role to fall back to permissioned
proposals and challenging (`PermissionedDisputeGame`) on Sepolia.

* fp: Enable permissionless dispute games task (#185)

Adds a task for the `Guardian` role to enable permissionless dispute
games (`FaultDisputeGame`) as the respected game type in the
`OptimismPortal`.

* Update tasks to work with deputy guardian module

* address review comments

* match chisel snippet with state diff example

---------

Co-authored-by: inphi <mlaw2501@gmail.com>
Co-authored-by: Maurelian <maurelian@protonmail.ch>