Fix DelayedWeth owner verification in MT Cannon task
## Description

This PR fixes the DelayedWeth owner verification in the SignFromJson.s.sol file for the MT Cannon task. The verification was previously disabled with a TODO comment indicating that the DelayedWeth owners were misconfigured.

## Changes

- Uncommented the owner verification check in the `checkWeth` function
- Removed the TODO comment as the issue is now fixed

## Testing

The change ensures that the system will properly validate that the owner of DelayedWeth instances matches the expected proxyAdminOwnerSafe address during verification.

## Related Issues

Addresses the TODO comment in the SignFromJson.s.sol file regarding misconfigured DelayedWeth owners.