Allow L1PAO to transfer weth contracts it is not the owner() of
Certain op chains were deployed with an incorrect configuration, such
that the owner of the DelayedWeth contracts was the SystemConfigOwner
rather than the L1PAO. This means that transferring ownership would
require two tasks if it were done solely using
`weth.transferOwnership().

To work around that issue, we modify this template so that it uses the
StorageSetter to set the owner slot.